feat: added code for the controllers

+++ services/blog.services.ts
@@ -1,6 +1,7 @@
 import BlogRepo from "../repository/blog.repo";
 import { BlogInterface } from "../types";
 import { BlogError } from "../error";
+import { StatusCodes } from "http-status-codes";
 
 const blogRepo = new BlogRepo();
 
@@ -14,9 +15,22 @@ export const createBlogService = async (blog: BlogInterface) => {
   return newBlog;
 };
 
-export const deleteBlogService = async (blogId: string) => {
+export const deleteBlogService = async (blogId: string, userId: string) => {
   if (!blogId) {
     throw new BlogError("blogId is required to delete a blog");
+  }
+
+  const blog = await blogRepo.getBlogById(blogId);
+
+  if (!blog) {
+    throw new BlogError("blog not found");
+  }
+
+  if (blog.author?.toString() !== userId) {
+    throw new BlogError(
+      "user is not authorized to delete this blog",
+      StatusCodes.UNAUTHORIZED
+    );
   }
 
   await blogRepo.deleteBlog(blogId);
@@ -74,7 +88,8 @@ export const getBlogByIdService = async (blogId: string) => {
 
 export const updateBlogService = async (
   blog: Omit<BlogInterface, "id">,
-  targetBlogId: string
+  targetBlogId: string,
+  userId: string
 ) => {
   if (!blog) {
     throw new BlogError("provide something to update");
@@ -84,25 +99,34 @@ export const updateBlogService = async (
     throw new BlogError("target blogId should be provided");
   }
 
+  const currentBlog = await blogRepo.getBlogById(targetBlogId);
+
+  if (currentBlog?.author?.toString() !== userId) {
+    throw new BlogError("user is not authorized to update this blog");
+  }
+
   const updatedBlog = blogRepo.updateBlog(blog, targetBlogId);
 
   return updatedBlog;
 };
 
-export const getTopBlogsService = async () => {
-  const limit = 5;
+export const getTopBlogsService = async (limit: number) => {
   const blogs = await blogRepo.getTopBlogs(limit);
   return blogs;
 };
 
-export const getBlogPageService = async (page: number) => {
+export const getBlogPageService = async (page: number, limit: number) => {
   if (page < 0) {
     throw new BlogError("page can never be less than zero");
   }
 
-  const limit = 5;
   const skip = (page - 1) * limit;
 
   const blogs = await blogRepo.getBlogPage(skip, limit);
   return blogs;
+};
+
+export const getTotalBlogsCountService = async () => {
+  const count = await blogRepo.getBlogCount();
+  return count;
 };
